cooking_agent/requirements.txt
agent-framework-azure-ai --pre
python-dotenv
openai
